220-nacos_client/pyproject.toml
[tool.poetry]
name = "220-nacos-client"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
nacos-sdk-python = "^0.1.14"
python-dotenv = "^1.0.0"


[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"
priority = "default"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
